CLI chat ( receiver side )
This the is program by with CLI chat can be done . The need to connect both users on same network . Here i design IP address for all systems ( common IP - 127.0.0.1). Used to send and receive msg self is system .Here receiver receives data send by the sender and store the data in txt document with time and date. And able to send data to the sender and (sender.py) sender  is able to receive data and store in txt document with date and time .